software_components/app_server/ampa_members_manager_project/requirements.txt
django==4.2
model-bakery==1.11.0
django-localflavor==3.1
schwifty==2023.3.0
phonenumbers==8.13.9
django-phonenumber-field[phonenumbers]==7.0.2
xlrd==2.0.1
django-extensions==3.2.1
xsdata==22.12
Unidecode==1.3.6
pandas==2.2.3
openpyxl==3.1.5
djangorestframework